lab4/requirements.txt
django==3.*
djangorestframework
mysqlclient